feat: move Message and LLMResponse to base.py for shared protocol types
Move shared dataclasses (Message, LLMResponse) from api_client.py to
base.py so they can be imported by all LLM provider implementations.
This establishes base.py as the canonical location for protocol types
that all providers must use.

Closes #73
